Add rooftop Agility, and check in the in-flight work across the tree
A single atomic commit of everything that was sitting unstaged in this shared
checkout, requested as one unit rather than split per feature.

Rooftop Agility (content/skills/agility, new):

The obstacle framework plus Draynor Village end to end. Two movement kinds cover
every rooftop obstacle in the game: `Movement.Cross` drives an `exactMove` (this
is the codebase's only forced-movement primitive, and had no content-side caller
until now) and `Movement.Climb` animates, delays and telejumps, which is
`LadderScript.climb` with a different destination. Obstacles pay their own
experience through `XpModifiers`; `LapProgress` tracks position in the chain and
silently restarts the lap on anything out of order, so only a clean lap earns the
completion bonus. Marks of grace are rolled once at the start of a lap and spawned
private to the player via `ObjRepository.add(receiver = ...)`. There is no failure
chance - every obstacle always succeeds.

All nine courses' obstacle locs are referenced; only Draynor's chain is authored.

Every rooftop obstacle turns out to carry its interaction on op1, without
exception, so the whole game binds with one `onOpLoc1`. That was verified by
decoding all 57,690 loc types straight from the cache rather than assumed - loc
ops are not uniform in this revision.

tools/agility (new): three standalone cache dumpers behind the coordinates above -
4.26M loc placements, every loc type, and 8.38M map tile settings, read without
booting a server. Landing tiles are not in the cache at all (OSRS hardcodes them),
so they are read off the decoded terrain, where rooftops show up as unblocked
corridors surrounded by blocked void. Two traps are documented in the README: rope
end markers sit on *blocked* terrain, and placement levels are raw and shift down
one level under LINK_BELOW.

`AgilityCourseTest` sweeps every authored destination against the real collision
map, which is what caught the rope-end case.

Also included, from other work in progress on this checkout: consumables drinks
and queues, the teleport menu and its destination table, generated drop-table
data, the farming skill module, first-login changes, the client build workflow
and packaging, the hiscores web API, and the deploy/client docs.

.gitignore now covers __pycache__ so the web tooling stops offering bytecode.

+++ content/custom/drop-tables/src/main/kotlin/org/rsmod/content/custom/droptables/DropTableRoller.kt
@@ -111,6 +111,11 @@ class DropTableRoller @Inject constructor(private val random: GameRandom) {
      * Shares [remainder] out across the empty slots in proportion to their authored weights, so an
      * unboosted table hands each one back exactly what it declared. Any rounding shortfall goes to
      * the last empty slot, keeping the weights summing to the roll space exactly.
+     *
+     * The share is computed in [Long] because the product overflows [Int] long before the quotient
+     * would: [remainder] reaches `denominator * scale` and a slot's weight reaches `denominator`,
+     * so a table any larger than `/6553` would wrap under a boosted roll. The quotient is always at
+     * most [remainder], so narrowing back is safe.
      */
     private fun distributeRemainder(
         slots: List<DropSlot>,
@@ -127,7 +132,7 @@ class DropTableRoller @Inject constructor(private val random: GameRandom) {
             if (slot !is DropSlot.Empty) {
                 continue
             }
-            val share = remainder * slot.weight / emptyWeightTotal
+            val share = (remainder.toLong() * slot.weight / emptyWeightTotal).toInt()
             weights[index] = share
             assigned += share
             lastEmpty = index
